Ejercicio 1 de Postman +

+++ workspace/monumentos/src/main/java/com/salesianostriana/dam/monumentos/repositories/MonumentoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MonumentoRepository extends JpaRepository<Monumento, Integer> {
+public interface MonumentoRepository extends JpaRepository<Monumento, Long> {
 
 }
